Update removeItem and clearAll functions to confirm from user before deleting data

+++ src/clearAll.js
@@ -1,10 +1,13 @@
 export default function clearAll() {
     let list = JSON.parse(localStorage.getItem('list'));
     document.getElementById('clearBtn').addEventListener('click', () =>{
-      list = list.filter((e) => !e.completed);
-      localStorage.setItem('list', JSON.stringify(list))
-      if (list !== null) {
-        location.reload();
+      const doubleCheck = confirm('Are you sure?');
+      if (doubleCheck) {
+        list = list.filter((e) => !e.completed);
+        localStorage.setItem('list', JSON.stringify(list))
+        if (list !== null) {
+          location.reload();
+        }
       }
     })
   }
