+++ src/removeItem.js
@@ -3,9 +3,12 @@ export default function removeItem() {
   const removeBtn = Object.values(document.getElementsByName('rmvbtn'));
   removeBtn.forEach((e) => {
     e.addEventListener('click', () => {
-      e.parentElement.remove();
-      list = list.filter((x) => !(e.id === x.id));
-      localStorage.setItem('list', JSON.stringify(list)); 
+      const doubleCheck = confirm('Are you sure?');
+      if (doubleCheck) {
+        e.parentElement.remove();
+        list = list.filter((x) => !(e.id === x.id));
+        localStorage.setItem('list', JSON.stringify(list)); 
+      }
       });
   });
 }
